\name{hcci-package}
\alias{hcci-package}
\alias{hcci}
\docType{package}
\title{Interval estimation for the parameters of linear models with heteroskedasticity of unknown form using bootstrap-t and percentile bootstrap and schemes of the double bootstrap.}
\description{This package calculates the interval estimates for the parameters of the models linares heteroskedasticity regression using bootstrap-t (Wild Bootstrap) and double bootstrap-t (Wild Bootstrap). It is also possible to calculate confidence intervals using the percentile bootstrap and percentile bootstrap double. It is also possible to calculate consistent estimates of the covariance matrix of the parameters of linear regression models with heteroskedasticity of unknown form. The package also provides function to calculate consistently the covariance matrix of the parameters of linear models with heteroskedasticity of unknown form.}
\details{
\tabular{ll}{
Package: \tab hcci\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2013-09-18\cr
License: \tab GPL (>=2)\cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Maintainer: Pedro Rafael Diniz Marinho <pedro.rafael.marinho@gmail.com>

Contributor: Francisco Cribari Neto <cribari@de.ufpe.br>
}
\references{
Cribari-Neto, F. (2004). Asymptotic inference under heteroskedasticity of unknown
form. Computational Statistics and Data Analysis, 45, 215-233.

Cribari-Neto, F.; Lima, M.G. (2009). Heteroskedasticity-consistent interval estimators. Journal of Statistical Computation and Simulation, 79, 787-803;

Cribari-Neto, F.; Souza, T.C.; Vasconcellos, K.L.P. (2007). Inference under heteros-
kedasticity and leveraged data. Communications in Statistics, Theory and Methods,
36, 1877-1888. [Errata: 37, 2008, 3329-3330.]

Horn, S.D.; Horn, R.A.; Duncan, D.B. (1975). Estimating heteroskedastic variances
in linear models. Journal of the American Statistical Association, 70, 380-385.

MacKinnon, J.G.; White, H. (1985). Some heteroskedasticity-consistent covariance
matrix estimators with improved finite-sample properties. Journal of Econometrics,
29, 305-325.

McCullough, B.D; Vinod, H.D. (1998). Implementing the double bootstrap, 12, 79-95.

White, H. (1980). A heteroskedasticity-consistent covariance matrix estimator and
a direct test for heteroskedasticity. Econometrica, 48, 817-838.

Wu, C.F.J. (1986). Jackknife, bootstrap and other resampling methods in regression analysis, 14, 1261-1295;

}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
%~~ documentation directory ~~
\keyword{heteroscedasticity}
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
