\name{schools}
\alias{schools}
\title{US Expenditures for Public Schools}
\description{
Per capita expenditure on public schools and per capita income
by state in 1979.
}
\usage{data(schools)}
\format{
A data frame containing 51 observations of 2 variables.
 \describe{
    \item{Expenditure}{per capita expenditure on public schools,}
    \item{Income}{per capita income.}
 }
}
\references{
Cribari-Neto F. (2004), Asymptotic Inference Under Heteroskedasticity
  of Unknown Form, Computational Statistics and Data Analysis, 45, 215-233.

Greene W.H. (1993), Econometric Analysis, 3nd edition. Macmillan Publishing Company, New York.

US Department of Commerce (1979), Statistical Abstract of the United States. US Government Printing Office, Washington, DC.
}

\keyword{datasets}
